temporary bomber animation

+++ Assets/Scripts/Player/Blaster.cs
@@ -9,7 +9,7 @@ public class Blaster : Weapon
 {
     public GameObject projectilePrefab;
     public float bulletSpd = 20;
-    public float Damage = 4;
+    public int Damage = 4;
     public float lifetime = 3;
     public int projectilesOnScreen = 2;
     public List<GameObject> currentProjectiles;
